+++ packages/components/__generated__/CommentItemRefetchQuery.graphql.ts
@@ -5,11 +5,12 @@
  */
 
 /* tslint:disable */
+
 /* eslint-disable */
 // @ts-nocheck
-
 import { ConcreteRequest, Query } from 'relay-runtime'
 import { FragmentRefs } from 'relay-runtime'
+
 export type CommentItemRefetchQuery$variables = {
   id: string
   isRepliesExpanded?: boolean | null | undefined