+++ packages/components/__generated__/CommentsList_comments.graphql.ts
@@ -5,11 +5,12 @@
  */
 
 /* tslint:disable */
+
 /* eslint-disable */
 // @ts-nocheck
-
 import { ReaderFragment, RefetchableFragment } from 'relay-runtime'
 import { FragmentRefs } from 'relay-runtime'
+
 export type CommentsList_comments$data = {
   readonly comments:
     | {