+++ packages/components/modules/navigations/NavigationLayout/MainContainer/types.ts
@@ -1,0 +1,13 @@
+import { PropsWithChildren } from 'react'
+
+import { BoxProps } from '@mui/material'
+
+export interface MainContainerProps extends PropsWithChildren {
+  isNavMini?: boolean
+  isNavHorizontal?: boolean
+  isNavCentered?: boolean
+}
+
+export interface NavVerticalContainerProps
+  extends BoxProps,
+    Pick<MainContainerProps, 'isNavMini'> {}