+++ packages/components/modules/navigations/shared/NavSectionHorizontal/NavList/NavItem/types.ts
@@ -1,0 +1,22 @@
+import { LinkProps } from '@mui/material'
+
+import {
+  NavItemProps as GenericNavItemProps,
+  NavItemBaseProps,
+  NavItemStateProps,
+} from '../../../../types'
+
+export interface StyledNavItemProps extends NavItemStateProps {
+  hasTabLayout?: boolean
+}
+
+export interface StyledLinkProps
+  extends LinkProps,
+    Pick<NavItemStateProps, 'active'>,
+    Pick<NavItemBaseProps, 'disabled'> {
+  hasTabLayout?: boolean
+}
+
+export interface NavItemProps extends GenericNavItemProps {
+  hasTabLayout?: boolean
+}