+++ packages/components/modules/navigations/shared/NavSectionVertical/Group/types.ts
@@ -1,0 +1,7 @@
+import type { NavItemBaseProps, SlotProps } from '../../../types'
+
+export type GroupProps = {
+  subheader?: string
+  items: NavItemBaseProps[]
+  slotProps?: SlotProps
+}