+++ packages/design-system/components/animate/index.ts
@@ -1,0 +1,4 @@
+export * from './variants'
+
+export { default as MotionContainer } from './MotionContainer'
+export { default as MotionViewport } from './MotionViewport'